Don't attempt to delete files which cant be deleted
Files uploaded via FTP, and thus processed via job templates, are marked as not deletable, causing the deletion attempt to mistakenly log an error

Fixes: OX-7431

+++ src/main/java/sirius/biz/storage/layer3/ExtractArchiveJob.java
@@ -88,7 +88,7 @@ public class ExtractArchiveJob extends SimpleBatchProcessJobFactory {
         this.overwriteExistingFilesParameter = new EnumParameter<>(OVERWRITE_EXISTING_FILES_PARAMETER_NAME,
                                                                    "$ExtractArchiveJob.overwriteExistingFilesParameter",
                                                                    ArchiveExtractor.OverrideMode.class).withDescription(
-                "$ExtractArchiveJob.overwriteExistingFilesParameter.help")
+                                                                                                               "$ExtractArchiveJob.overwriteExistingFilesParameter.help")
                                                                                                        .withDefault(
                                                                                                                ArchiveExtractor.OverrideMode.ON_CHANGE)
                                                                                                        .build();
@@ -105,7 +105,7 @@ public class ExtractArchiveJob extends SimpleBatchProcessJobFactory {
         if (sourceParameter == null) {
             sourceParameter = new FileParameter("source",
                                                 "$ExtractArchiveJob.sourceParameter").withAcceptedExtensionsList(new ArrayList<>(
-                    extractor.getSupportedFileExtensions()))
+                                                                                             extractor.getSupportedFileExtensions()))
                                                                                      .withDescription(
                                                                                              "$ExtractArchiveJob.sourceParameter.help")
                                                                                      .markRequired()
@@ -140,7 +140,9 @@ public class ExtractArchiveJob extends SimpleBatchProcessJobFactory {
 
         process.forceUpdateState(NLS.get("ExtractArchiveJob.completed"));
 
-        if (!process.isErroneous() && process.require(deleteArchiveParameter).booleanValue()) {
+        if (!process.isErroneous()
+            && process.require(deleteArchiveParameter).booleanValue()
+            && sourceFile.canDelete()) {
             process.log(ProcessLog.info().withNLSKey("ExtractArchiveJob.deletingArchive"));
             sourceFile.delete();
         }
